fix
Implemented PatientId as int across the service.
Changed domain entities, repositories, DTOs, integration events, application services, API route constraints, infrastructure repositories, factory context, and test doubles/BDD bindings. Kept Reminder.Id, EntityId, NotificationLog IDs, GUIDs, JWT, and [Authorize] untouched.

+++ src/MediTrack.ReminderService.Application/Dtos/NotificationPreferenceDto.cs
@@ -4,7 +4,7 @@ namespace MediTrack.ReminderService.Application.Dtos;
 
 /// <summary>Vista de lectura de las preferencias de notificación de un paciente.</summary>
 public sealed record NotificationPreferenceDto(
-    long PatientId,
+    int PatientId,
     bool SoundEnabled,
     bool VibrationEnabled,
     int RepeatCount,
